Running with arguments
